nasdaq sectors/datacamp-nasdaq-openai/requirements.txt
pandas>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0
